ZNTA-2022 - seperate post commit event listener
